circle/pyproject.toml
[tool.poetry]
name = "circle"
version = "0.1.0"
description = ""
authors = ["Sway Ducky <swayducky@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.9"
pillow = "^9.5.0"
flask = "^2.3.2"
flask-cors = "^3.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
